// wink-plugin-peripherals/builtin/seg_display/1.0.0/src/variants.generated.ts
// [AUTO-GENERATED by @wink-ai/unisim-ui/generator. DO NOT EDIT MANUALLY.]
// Source: wokwi-7segment
import type { PinsOverlayMap } from '@wink-ai/unisim-ui';

export const SEG_DISPLAY_AUTOGEN_SIZE = Object.freeze({
  width: 210,
  height: 96,
});

export const SEG_DISPLAY_AUTOGEN_OVERLAY: Readonly<PinsOverlayMap> = Object.freeze({
  A: Object.freeze({
    relX: 23,
    relY: 96,
    wireNet: 'primary' as const,
    defaultConnection: null,
    required: false,
  }),
  B: Object.freeze({
    relX: 47,
    relY: 96,
    wireNet: 'primary' as const,
    defaultConnection: null,
    required: false,
  }),
  C: Object.freeze({
    relX: 70,
    relY: 96,
    wireNet: 'primary' as const,
    defaultConnection: null,
    required: false,
  }),
  D: Object.freeze({
    relX: 93,
    relY: 96,
    wireNet: 'primary' as const,
    defaultConnection: null,
    required: false,
  }),
  E: Object.freeze({
    relX: 117,
    relY: 96,
    wireNet: 'primary' as const,
    defaultConnection: null,
    required: false,
  }),
  F: Object.freeze({
    relX: 140,
    relY: 96,
    wireNet: 'primary' as const,
    defaultConnection: null,
    required: false,
  }),
  G: Object.freeze({
    relX: 163,
    relY: 96,
    wireNet: 'primary' as const,
    defaultConnection: null,
    required: false,
  }),
  DP: Object.freeze({
    relX: 187,
    relY: 96,
    wireNet: 'primary' as const,
    defaultConnection: null,
    required: false,
  }),
});